+++ src/modules/menu/i18n/index.js
@@ -1,4 +1,4 @@
-import { provide as toolboxProvicer } from './toolbox.provider';
+import { provide as menuProvider } from './menu.provider';
 import { $injector } from '../../../injection';
 const { TranslationService: translationService } = $injector.inject('TranslationService');
-translationService.register('toolboxProvider', toolboxProvicer);
+translationService.register('toolboxProvider', menuProvider);
